+++ Proyecto/.idea/workspace.xml
@@ -1,16 +1,40 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
   <component name="ChangeListManager">
     <list default="true" id="f6df1ae7-b934-4f30-a0fa-9569583274a1" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/compilation/ast/auxiliary.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/ast/assignments.py" beforeDir="false" afterPath="$PROJECT_DIR$/compilation/ast/assignments.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/ast/initializations.py" beforeDir="false" afterPath="$PROJECT_DIR$/compilation/ast/initializations.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/ast/nodes.py" beforeDir="false" afterPath="$PROJECT_DIR$/compilation/ast/nodes.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/ast/nodes_yy.py" beforeDir="false" afterPath="$PROJECT_DIR$/compilation/ast/complex.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/ast/nodes_zz.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/ast/operations.py" beforeDir="false" afterPath="$PROJECT_DIR$/compilation/ast/operations.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/ast/parser_xx.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/enums.py" beforeDir="false" afterPath="$PROJECT_DIR$/compilation/enums.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/compilation/parser.py" beforeDir="false" afterPath="$PROJECT_DIR$/compilation/parser.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/compilation/variables.py" beforeDir="false" afterPath="$PROJECT_DIR$/compilation/variables.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
     <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
     <option name="LAST_RESOLUTION" value="IGNORE" />
   </component>
+  <component name="FileTemplateManagerImpl">
+    <option name="RECENT_TEMPLATES">
+      <list>
+        <option value="Python Script" />
+      </list>
+    </option>
+  </component>
   <component name="MarkdownSettingsMigration">
     <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProblemsViewState">
+    <option name="selectedTabId" value="ProjectErrors" />
   </component>
   <component name="ProjectId" id="226XBl1uHbE1PL3yh7JWZpq1ldx" />
   <component name="ProjectLevelVcsManager" settingsEditedManually="true" />
@@ -23,8 +47,12 @@
     <property name="RunOnceActivity.ShowReadmeOnStart" value="true" />
     <property name="WebServerToolWindowFactoryState" value="false" />
     <property name="last_opened_file_path" value="$USER_HOME$/AppData/Local/Programs/Python/Python37/python.exe" />
+    <property name="node.js.detected.package.eslint" value="true" />
+    <property name="node.js.selected.package.eslint" value="(autodetect)" />
+    <property name="node.js.selected.package.tslint" value="(autodetect)" />
     <property name="nodejs_interpreter_path.stuck_in_default_project" value="undefined stuck path" />
     <property name="nodejs_npm_path_reset_for_default_project" value="true" />
+    <property name="run.code.analysis.last.selected.profile" value="pProject Default" />
     <property name="settings.editor.selected.configurable" value="preferences.general" />
   </component>
   <component name="RunManager" selected="Python.main">
@@ -83,14 +111,32 @@
       <workItem from="1639158878497" duration="3251000" />
       <workItem from="1639243313222" duration="58000" />
       <workItem from="1639243390036" duration="8523000" />
+      <workItem from="1642607975678" duration="919000" />
+      <workItem from="1642617932196" duration="37000" />
+      <workItem from="1643126169808" duration="2427000" />
     </task>
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
     <option name="version" value="3" />
   </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
+  </component>
   <component name="com.intellij.coverage.CoverageDataManagerImpl">
-    <SUITE FILE_PATH="coverage/Proyecto$main.coverage" NAME="main Coverage Results" MODIFIED="1639161480649" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$" />
+    <SUITE FILE_PATH="coverage/Proyecto$test_nodes.coverage" NAME="test_nodes Coverage Results" MODIFIED="1639959578342" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/compilation/ast" />
+    <SUITE FILE_PATH="coverage/Proyecto$misano.coverage" NAME="misano Coverage Results" MODIFIED="1642353701852" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$USER_HOME$/Downloads/Telegram Desktop" />
+    <SUITE FILE_PATH="coverage/Proyecto$assignments.coverage" NAME="assignments Coverage Results" MODIFIED="1639698156044" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/compilation" />
+    <SUITE FILE_PATH="coverage/Proyecto$start.coverage" NAME="start Coverage Results" MODIFIED="1639701394053" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$" />
+    <SUITE FILE_PATH="coverage/Proyecto$main.coverage" NAME="main Coverage Results" MODIFIED="1642468471116" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$" />
     <SUITE FILE_PATH="coverage/Proyecto$.coverage" NAME=" Coverage Results" MODIFIED="1639159054883" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="true" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$/tester" />
   </component>
 </project>
